modified:   client/src/components/Login.js     Removed the modalKey from Modal modified:   client/src/components/Modal.js     Removed useModalRegistration, the registration should be when the component is been called from renamed:    client/src/components/Modal/actions.js -> client/src/components/Modal/modalActons.js     Change file name to match what the file is intended todo modified:   client/src/components/Modal/modalReducer.js     added a deregister option to remove the modal from registration modified:   client/src/layouts/components/TopNavbar.js     Removed the modalKey renamed:    client/src/utils/modalActions.js -> client/src/utils/modalHooks.js     Remanded the file to match what the file is intended todo
