+++ client/src/components/Modal/modalActons.js
@@ -1,4 +1,5 @@
 export const REGISTER_MODAL = "REGISTER_MODAL";
 export const SET_ACTIVE_MODAL = "SET_ACTIVE_MODAL";
+export const DEREGISTER_MODAL = "DEREGISTER_MODAL";
