+++ client/src/components/Modal/modalReducer.js
@@ -1,21 +1,24 @@
-import { SET_ACTIVE_MODAL, REGISTER_MODAL } from "./actions.js";
+import { SET_ACTIVE_MODAL, REGISTER_MODAL, DEREGISTER_MODAL } from "./modalActons.js";
 
 const reducer = (state, { type, payload }) => {
 	const actions = {
-		[REGISTER_MODAL]: ()=>(
-			{
-				...state,
-				modals: { ...state.modals, ...payload }
+		[REGISTER_MODAL]: () => ({
+			...state,
+			modals: { ...state.modals, ...payload }
+		}),
+		[SET_ACTIVE_MODAL]: () => ({
+			...state,
+			activeKey: payload
+		}),
+		[DEREGISTER_MODAL]: () => ({
+			...state,
+			modals: {
+				...Object.fromEntries(
+					Object.entries(state.modals).filter(([key, value]) => key !== payload)
+					)
 			}
-		),
-		[SET_ACTIVE_MODAL]: ()=>(
-				{
-				...state,
-				activeKey: payload
-			}
-		)
-			
-	} 
+		})
+	}; 
 	return actions.hasOwnProperty(type) ? actions[type]() : state;
 
 };
